refactor(memory-wiki): share person-page classification
